Move all embulk-junit4 classes to org.embulk.test
