+++ embulk-core/src/test/java/org/embulk/plugin/jar/TestJarPluginLoader.java
@@ -12,8 +12,8 @@ import java.util.Collections;
 import java.util.concurrent.Callable;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
-import org.embulk.EmbulkTestRuntime;
 import org.embulk.plugin.PluginClassLoader;
+import org.embulk.test.EmbulkTestRuntime;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
