+++ embulk-core/src/test/java/org/embulk/spi/TestFileInputRunner.java
@@ -8,13 +8,14 @@ import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Queue;
-import org.embulk.EmbulkTestRuntime;
 import org.embulk.config.ConfigDiff;
 import org.embulk.config.ConfigSource;
 import org.embulk.config.TaskReport;
 import org.embulk.config.TaskSource;
-import org.embulk.spi.TestPageBuilderReader.MockPageOutput;
 import org.embulk.spi.util.Pages;
+import org.embulk.test.EmbulkTestRuntime;
+import org.embulk.test.MockParserPlugin;
+import org.embulk.test.TestPageBuilderReader.MockPageOutput;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
