+++ embulk-core/src/test/java/org/embulk/spi/TestFileOutputRunner.java
@@ -9,12 +9,15 @@ import static org.msgpack.value.ValueFactory.newString;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
-import org.embulk.EmbulkTestRuntime;
 import org.embulk.config.ConfigDiff;
 import org.embulk.config.ConfigSource;
 import org.embulk.config.Task;
 import org.embulk.config.TaskReport;
 import org.embulk.config.TaskSource;
+import org.embulk.test.EmbulkTestRuntime;
+import org.embulk.test.MockFormatterPlugin;
+import org.embulk.test.MockParserPlugin;
+import org.embulk.test.PageTestUtils;
 import org.junit.Rule;
 import org.junit.Test;
 import org.msgpack.value.ImmutableMapValue;
