+++ embulk-junit4/src/main/java/org/embulk/test/EmbulkTestRuntime.java
@@ -1,6 +1,12 @@
-package org.embulk;
+package org.embulk.test;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.UncheckedIOException;
+import java.util.Collections;
+import java.util.Properties;
 import java.util.Random;
+import java.util.Set;
 import org.embulk.EmbulkEmbed;
 import org.embulk.config.ModelManager;
 import org.embulk.exec.PooledBufferAllocator;
@@ -11,8 +17,6 @@ import org.embulk.spi.BufferAllocator;
 import org.embulk.spi.ExecAction;
 import org.embulk.spi.ExecInternal;
 import org.embulk.spi.ExecSessionInternal;
-import org.embulk.spi.MockFormatterPlugin;
-import org.embulk.spi.MockParserPlugin;
 import org.junit.rules.TestRule;
 import org.junit.rules.TestWatcher;
 import org.junit.runner.Description;
@@ -57,7 +61,7 @@ public class EmbulkTestRuntime implements TestRule {
     }
 
     public static PluginClassLoaderFactory buildPluginClassLoaderFactory() {
-        return PluginClassLoaderFactoryImpl.of(EmbulkEmbed.PARENT_FIRST_PACKAGES, EmbulkEmbed.PARENT_FIRST_RESOURCES);
+        return PluginClassLoaderFactoryImpl.of(PARENT_FIRST_PACKAGES, PARENT_FIRST_RESOURCES);
     }
 
     @Override
@@ -102,4 +106,25 @@ public class EmbulkTestRuntime implements TestRule {
     private static org.embulk.config.ModelManager createModelManager() {
         return new org.embulk.config.ModelManager();
     }
+
+    private static Set<String> readPropertyKeys(final String name) {
+        try (final InputStream in = EmbulkEmbed.class.getResourceAsStream(name)) {
+            if (in == null) {
+                throw new NullPointerException(
+                        String.format("Resource '%s' is not found in classpath. Jar file or classloader is broken.", name));
+            }
+            final Properties properties = new Properties();
+            properties.load(in);
+            return Collections.unmodifiableSet(properties.stringPropertyNames());
+        } catch (final IOException ex) {
+            throw new UncheckedIOException(ex);
+        }
+    }
+
+    // EmbulkEmbed.PARENT_FIRST_PACKAGES and EmbulkEmbed.PARENT_FIRST_RESOURCES are package-private.
+    // They are not accessible from org.embulk.test.
+    //
+    // TODO: Remove them finally.
+    private static final Set<String> PARENT_FIRST_PACKAGES = readPropertyKeys("/embulk/parent_first_packages.properties");
+    private static final Set<String> PARENT_FIRST_RESOURCES = readPropertyKeys("/embulk/parent_first_resources.properties");
 }
